+++ tests/test_pretty.py
@@ -1,3 +1,5 @@
+from array import array
+from collections import defaultdict
 import io
 import sys
 
@@ -17,21 +19,24 @@ def test_install():
 def test_pretty():
     test = {
         "foo": [1, 2, 3, {4, 5, 6, (7, 8, 9)}, {}],
+        "bar": {"egg": "baz", "words": ["Hello World"] * 10},
         False: "foo",
         True: "",
         "text": ("Hello World", "foo bar baz egg"),
     }
 
-    result = pretty_repr(test)
-    expected = "{\n    'foo': [1, 2, 3, {(7, 8, 9), 4, 5, 6}, {}], \n    False: 'foo', \n    True: '', \n    'text': ('Hello World', 'foo bar baz egg')\n}"
-    assert result.plain == expected
+    result = pretty_repr(test, max_width=80)
+    print(result)
+    print(repr(result))
+    expected = "{\n    'foo': [1, 2, 3, {(7, 8, 9), 4, 5, 6}, {}],\n    'bar': {\n        'egg': 'baz',\n        'words': [\n            'Hello World',\n            'Hello World',\n            'Hello World',\n            'Hello World',\n            'Hello World',\n            'Hello World',\n            'Hello World',\n            'Hello World',\n            'Hello World',\n            'Hello World'\n        ]\n    },\n    False: 'foo',\n    True: '',\n    'text': ('Hello World', 'foo bar baz egg')\n}"
+    assert result == expected
 
 
 def test_small_width():
     test = ["Hello world! 12345"]
     result = pretty_repr(test, max_width=10)
     expected = "[\n    'Hello world! 12345'\n]"
-    assert result.plain == expected
+    assert result == expected
 
 
 def test_broken_repr():
@@ -41,8 +46,8 @@ def test_broken_repr():
 
     test = [BrokenRepr()]
     result = pretty_repr(test)
-    expected = "[<error in repr: division by zero>]"
-    assert result.plain == expected
+    expected = "[<repr-error 'division by zero'>]"
+    assert result == expected
 
 
 def test_recursive():
@@ -50,4 +55,20 @@ def test_recursive():
     test.append(test)
     result = pretty_repr(test)
     expected = "[...]"
-    assert result.plain == expected
+    assert result == expected
+
+
+def test_defaultdict():
+    test_dict = defaultdict(int, {"foo": 2})
+    result = pretty_repr(test_dict)
+    assert result == "defaultdict(<class 'int'>, {'foo': 2})"
+
+
+def test_array():
+    test_array = array("I", [1, 2, 3])
+    result = pretty_repr(test_array)
+    assert result == "array('I', [1, 2, 3])"
+
+
+def test_tuple_of_one():
+    assert pretty_repr((1,)) == "(1,)"