+++ tools/stress_test_pretty.py
@@ -3,7 +3,7 @@ from rich.panel import Panel
 from rich.pretty import Pretty
 
 DATA = {
-    "foo": [1, 2, 3, (1, 2, 3), {4, 5, 6, (7, 8, 9)}, "Hello, World"],
+    "foo": [1, 2, 3, (), {}, (1, 2, 3), {4, 5, 6, (7, 8, 9)}, "Hello, World"],
     "bar": [None, (False, True)] * 2,
 }
 console = Console()